refactor(react-tinacms-inline): extract img components to own files
